controller: fix CM OwnRef on update
When updating our CM with the new IP we were passing in the wrong UID.
Instead of passing in the UID of our deployment we were passing in
the UID of ourselves to serve as a Owner Reference. What this caused was
that our CM was not destroyed when deleting our deployment.