Check dependencies are present in Pod handler
If either the Deployment or the ConfigMap are not present, the Pod
handler should exit early.

This happens because of the different rate at which resources are
deleted. If a SG is deleted, the Deployment and the ConfigMap are as
well.

But the Pod might take a little longer to be deleted, and in the
meantime, events for it will still be fired.

The handler should therefore check whether the resources it needs are
present, and if they aren't, it should simply exit.