*: introduce initial configuration
user.toml file serves as a way of initial configuration in habitat. Our
operator mounts this file in a Volume as a Secret that a user has to
create first. Habitat will pick up this file and all the
configurations in this file will override the ones defined in
default.toml.