*: support Habitat bind feature
Runtime binding in Habitat refers to the ability for one service group
to connect to another forming a producer/consumer relationship where the
consumer can use the producer's publicly available configuration to
configure it's services at runtime.

Because bind can be specified multiple times, we added support for this
feature by reading it as an array from the yaml manifest.