controller: Fix updating user.toml
Revert the SubPath usage and change the destination where we
bind-mount a directory. The new destination is supposed to contain
only the user.toml file, so hiding the previous contents of the
directory should not be a problem anymore. Not using SubPath works
around the atomic updates on bind-mounted file problems.