Add channels support to the Habitat type
Habitat applications can exist in the different "channels" -
like "staging", "production", "unstable" etc. Those names
can be configured by the user.

Those channels will be expressed as labels in Kubernetes.

There is also a possibiity of promoting Habitat applications.
This action is expressed by HabitatPromote object, which needs
to be created in order to perform it.

To test that functionality, you can use the example yml files:

    kubectl create -f examples/standalone/habitat.yml
    kubectl create -f examples/standalone/habitat-promotion.yml

Fixes #126

Signed-off-by: Michal Rostecki <michal@kinvolk.io>