Refactor e2e tests
Due to the long time the e2e tests were running, a refactor was needed.
This included removing the tests and testing everything under the
bind-config test. This tests includes creation, updating, binding of
Habitat services so it is a true end-to-end test.

With this commit we also take the resources from a manifest file rather
then creating the resources manually.