Add Namespace to new ConfigMap
ConfigMaps are saved to the cache with a key that looks like
`namespace/name`.

The ConfigMap objects created by `newConfigMap` though were not
inheriting the Habitat's namespace, which means that when they were
passed to MetaNamespaceIndexFunc they would return a key without
namespace.

This means that we were using one key to save, and a different key to
retrieve, leading to ConfigMaps not being found.

Signed-off-by: Lorenzo Manacorda <lorenzo@kinvolk.io>