Update the supported Habitat version
The oldest version of Habitat that operator can use is 0.39, because
this is the first version that has the alternative path for
user.toml. But Habitat 0.52 is the first version that watches for
changes in user.toml if it is placed in the alternative path.

So for simplicity, let's just say we support Habitat version from 0.52
onwards.