Remove Headless Service
We don't need to have a Headless Service to have the Pods find each other.
They use the gossip protocol for that.

I decided against simply removing the commits that add the Headless
Service, since we might want to add it again in the future (e.g. to
replace the current ring-joining solution based on ConfigMaps and file
watching), and creating it through the API is a relatively tedious and
error-prone process, so having this in the code might save us some time
and frustrations.

Signed-off-by: Lorenzo Manacorda <lorenzo@kinvolk.io>