Make RBAC in tests namespace-bound
There is no need to grant cluster-level access to the operator in E2E
tests, as they all get executed within the testing namespace.

This will allow us to more easily clean up after the tests by simply
deleting the whole namespace, a change that will be performed
in a following patch.

Signed-off-by: Lorenzo Manacorda <lorenzo@kinvolk.io>