Make it possible to use sudo when calling docker
Docker on Fedora requires root to talk to it, so `make image` always
fails there currently. But let's not just hardcode `sudo` to fix it,
because other distros may be more convenient here by not requiring
root when talking to docker, so sudo here is pointless. With this
change, we can do `make SUDO=sudo image` on Fedora to make it work.