Time is hard
Check for the timeout in ticker case too, because go does not
guarantee the ordering of the evaluation of the select cases, so it is
possible that by the time we reach the select, both ticker and timer
have data in their channels, but only the ticker case will be
evaluated. In this case check the timer channel too in a nonblocking
manner.