stateful sets: Set UpdateStrategy to RollingUpdate
The default value is supposed to be `RollingUpdate` if unspecified but
this doesn't seem to be the case with `v1beta1` at the moment. Also
updating to `v1beta2` is not possible since this breaks support for
Kubernetes 1.7.0.

We should ideally move to `v1` once we support Kubernetes 1.10 and
then can drop support for 1.7.

See issue #235 for more details.