Fix: Use correct mongodb image
In order for this example to work, the application needs to have access
to a mongo instance with the National Parks data.  This is provided by
the `habitat/np-mongodb` plan, currently maintained by CFT.

This includes a post-run hook that drops the demo database, creates a
nationalparks database, and runs a mongo import using JSON which
specifies the location of the parks.

There is some question over whether a sleep is needed in the post-run
hook to ensure that Tomcat has come up in time.  This smells fishy, and
should be looked at.

Signed-off-by: Stephen Nelson-Smith <stephen.nelsonsmith@ticketmaster.co.uk>