RBAC: Allow operator to manage both deployments and statefulsets
The example RBAC file wasn't updated after adding a custom CRD
versions support - it allowed the operator to manage only statefulsets
as at the point of the change the operator handled only them;
deployments were dropped. Right now the operator can manage both
deployments and statefulsets depending on the custom version specified
in the Habitat resource manifest.

The manifest file for the helm chart was also out-of-sync with the
changes in the operator - the rules allowed operator only to manage
deployments.

Signed-off-by: Krzesimir Nowak <krzesimir@kinvolk.io>