CHANGELOG: Deprecate customVersion = v1beta1
We follow the release schedule of Kubernetes, according to which
deprecated versions are dropped after 3 releases.

`apps/v1beta2` was [introduced][v1beta2] in 1.8, thereby deprecating `v1beta1`, which
should therefore be dropped by 1.11.

`customVersion = v1beta1` maps to a controller that uses `apps/v1beta1`,
which won't be updated, and that must therefore be droppped.

[v1beta2]: https://github.com/kubernetes/kubernetes/blob/master/CHANGELOG-1.8.md#workloads-api-appsv1beta2

Signed-off-by: Lorenzo Manacorda <lorenzo@kinvolk.io>