examples: Use labels
We need to improve the naming of our example Habitats.
At the moment, we sometimes use example-x-y-z.
This is fine in some cases (example-standalone-habitat), but then we also use "go" in other cases.
From a discoverability/UX point of view, using labels would free us to use the name that makes the most sense,
but still letting the user clearly know that the Habitat is from the operator and is an example.

Signed-off-by: Kosy Anyanwu <kosy@kinvolk.io>