controller: Write multiple IP to peer-watch-file
Problem: While implementing multiple instances in the
habitat-service-broker we noticed that the supervisors get stuck in
the leader election phase when applying a service binding. This
happens as the StatefulSet uses a RollingUpdate strategy for
updating the deployment. As a result, while each pod is updated at a
time, it appears at the moment that the supervisors running in the
pods have a stale copy of the peer-watch-file. How this happens is not
very clear at the moment, but our understanding is that this may be
the result of the supervisors not listening for changes to the
peer-watch-file once they have joined the ring. This can be fixed in
the habitat supervisor but would likely require additional brain
storming.

At the moment, we write only one IP address from the list of
running pods to the peer-watch-file. This increases the likelihood
that if the pod whose IP address is in the peer-watch-file is deleted,
other supervisors would still have the stale copy of the
peer-watch-file as they have already joined the ring before this pod
was deleted and have stopped listening for changes to the file.

Solution: (More like a workaround) We are changing this approach by
writing multiple pod IP addresses to the peer-watch-file instead of a
single one. This was found to have fixed the bug where the supervisors
get stuck in the leader election phase as even if a pod is deleted,
there are other pods from the list in peer-watch-file actively
running.

Food for thought: In our tests with minikube, it was found that deleted pods
once rescheduled were allocated the same IP address as the one it had
before being deleted, provided no other pods were started or deleted
in the meantime. While this is not guaranteed, it was expected that
the supervisors would have resumed operation without getting stuck as
rescheduling a pod and ending up with the same IP would mean nothing
changed in the network from the perspective of the supervisors already
running. This might warrant a closer look at the ring
formation/joining and leader election of the supervisors.

Signed-off-by: Indradhanush Gupta <indra@kinvolk.io>