controller: Delete all pods on update to Sts
A supervisor bug prevents pods from successfully completing a leader
election when an update is made to the StatefulSet object. This arises
from the RollingUpdate behaviour of StatefulSet.

However, deleting them all at once fixes the problem. In this commit
we compare the older StatefulSet object with the updated one and
delete all pods that belong to that object if the StatefulSet objects
before and after the update are different. This check was necessary
because the StatefulSet keeps getting updated even though nothing has
changed as a result of the ConfigMap cache getting updated frequently.

More about the bug here: habitat-sh/habitat#5264

Signed-off-by: Indradhanush Gupta <indra@kinvolk.io>