ci: Skip steps if required tools were not set up
Some of the steps are always executed or are executed on
failure. Those steps require some things to be set up. But if some
test fails before those things are set up, then we get spurious error
messages from these steps too. This may happen when RBAC rules are not
in sync, so the tests will fail before setting up the cluster and
pushing the docker image to container registry, so there will be no
cluster to get the logs from (even likely there won't be even a
kubectl binary to use) and to delete nor images to remove from the
registry.

Signed-off-by: Krzesimir Nowak <krzesimir@kinvolk.io>