[version] Display version when booting operator
Operator shows it's version when it starts. This version information
is embedded inside the binary while building it. It is a good
practice to show the operator version when it is booting up, helps
while resolving/reporting issues even if the latest/untagged docker
image is run by the user.

Signed-off-by: Suraj Deshmukh <suraj@kinvolk.io>