[RBAC] Harden RBAC policies for the operator
Right now we ask user to create ClusterRole and ClusterRoleBinding.
Which grants the operator cluster wide permissions, which is too
much of permissions.

So this PR tries to limit those permissions by assigning only Role
and RoleBinding, which are namespace limited.

* Add new flags `namespace` where you can define which namespace
this operator should run inside.

* Add a check to the code if no namespace is provided whether the
user has rights to list the namespaces, if not then the operator
fails.

* Add flag `assumeCRDRegistered` which is way to tell the operator
whether CRD has already been created inside the cluster by admin.

Signed-off-by: Suraj Deshmukh <suraj@kinvolk.io>