[test] sync RBAC rules in various places
* Add golang test to sync Roles and ClusterRoles in various places in
the code repository. These manifests exists in examples, tests and
helm charts. If anyone is changed this test enforces all of them to
change, this ensures that no manifest is outdated with changes in
permissions.

* This checks synchronization of Roles in various places.

* This checks synchronization of ClusterRoles in various places.

* This also checks if the Roles and ClusterRoles are in sync.
ClusterRoles have only two extra permissions than Roles viz.
Namespaces and CRDs, excluding those everything else is in same. So
excluding those two permissions this test checks if other permissions
are in sync.

* Run this as part of `test` target of Makefile.

* Remove the existing shell script that used to check the
synchronization of these manifests.

Signed-off-by: Suraj Deshmukh <suraj@kinvolk.io>